+++ keyAniManager.h
@@ -29,9 +29,9 @@ public:
 	void addDefaultFrameAnimation(string animationKeyName, const char* imageKeyName, int fps, bool reverse, bool loop, void* cbFunction, void* obj);
 	
 	//배열 애니메이션
-	void addArrayFrameAnimation(string animationKeyName, const char* imageKeyName, int* arr, int arrLen, int fps, bool loop);
-	void addArrayFrameAnimation(string animationKeyName, const char* imageKeyName, int* arr, int arrLen, int fps, bool loop, void* cbFunction);
-	void addArrayFrameAnimation(string animationKeyName, const char* imageKeyName, int* arr, int arrLen, int fps, bool loop, void* cbFunction, void* obj);
+	void addArrayFrameAnimation(string animationKeyName, const char* imageKeyName, int* arr, int arrLen, float fps, bool loop);
+	void addArrayFrameAnimation(string animationKeyName, const char* imageKeyName, int* arr, int arrLen, float fps, bool loop, void* cbFunction);
+	void addArrayFrameAnimation(string animationKeyName, const char* imageKeyName, int* arr, int arrLen, float fps, bool loop, void* cbFunction, void* obj);
 	
 	//구간 애니메이션
 	void addCoordinateFrameAnimation(string animationKeyName, const char* imageKeyName, int start, int end, int fps, bool reverse, bool loop);
